requirements.txt
amrlib==0.8.0
coloredlogs==15.0.1
datasets==3.5.0
diskcache==5.6.3
matplotlib==3.9.4
mlflow==2.22.0
networkx==3.2.1
nevergrad==1.0.12
numpy==2.0.2
pandas==2.2.3
Penman==1.3.1
psutil==7.0.0
PyYAML==6.0.2
ray==2.44.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==4.1.0
spacy==3.8.5
stanza==1.10.1
torch==2.3.1+cu121
torch-geometric==2.6.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.3.1+cu121
torchtext==0.18.0
torchvision==0.18.1+cu121
tqdm==4.67.1
transformers==4.51.3
